requirements.txt
python-dotenv
requests
openai
streamlit
pymongo